openlp: update to 2.4.6 & add missing rundeps
Summary:
Due to some differences between the various platforms we support, and some bug fixes for some of those platforms, OpenLP 2.4.5 unfortunately introduced some bugs. Thankfully they were easy enough to pick up, and easy enough to fix.

In this bugfix release we also fixed a problem where the OpenLP database importer was dropping the author type, and a problem with some of the core translations not being loaded.

Here is the complete list of bugs we fixed:

* Fixed a bug where the author type upgrade was being ignored because it was looking at the wrong table
* Fixed a bug where the songs_songbooks table was not being created because the if expression was the wrong way round
* Changed the songs_songbooks migration SQL slightly to take into account a bug that has (hopefully) been fixed
* Sometimes the timer goes off as OpenLP is shutting down, and the application has already been deleted
* Fixed opening the data folder (KDE thought the old way was an SMB share)
* Fixed a problem with the new QMediaPlayer not controlling the playlist anymore
* Added importing of author types to the OpenLP 2 song importer
* Fixed a problem with loading Qt's translation files
* Disabled the controls in the shortcut dialog unless a shortcut is actually selected

Test Plan:
Ran the program, consulted wiki to verify rundeps, went through setup wizard,
played around with some of the options.

Reviewers: #triage_team, sunnyflunk

Reviewed By: #triage_team, sunnyflunk

Subscribers: sunnyflunk

Maniphest Tasks: T3932

Differential Revision: https://dev.solus-project.com/D465